Added minimum step size for infiniteScroll tableType
