+++ src/app/pages/home/home.page.html
@@ -11,7 +11,7 @@
         [collRef]=""
         [columns]="columns" 
         [snapshot]="false"
-        [steps]="10" 
+        [steps]="5" 
         [displayOptions]="{ 
             tableType: 'infiniteScroll', 
             fixed: true
