+++ src/app/shared/components/tables/table-configurable/table-configurable.component.html
@@ -71,7 +71,7 @@
 <ng-template #infiniteScroll>
 	<ion-infinite-scroll
 		(ionInfinite)="handleScroll($event)"
-		threshold="100px"
+		threshold="200px"
 		[disabled]="disableInfiniteScroll"
 	>
 		<ion-infinite-scroll-content
